Migrate LinearBlend and SingleState update logic to static class
