+++ Runtime/Utils/ClipSamplerUtils.cs
@@ -12,17 +12,19 @@ namespace DMotion
         internal const int MaxSamplersCount = byte.MaxValue / 2;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool AddWithId<T>(this DynamicBuffer<T> samplers, T newSampler, out byte id, out int index) where T : struct, IElementWithId
+        internal static bool AddWithId<T>(this DynamicBuffer<T> samplers, T newSampler, out byte id, out int index)
+            where T : struct, IElementWithId
         {
-             if (samplers.TryFindIdAndInsertIndex(1, out id, out index))
-             {
-                 newSampler.Id = id;
-                 samplers.Insert(index, newSampler);
-                 return true;
-             }
-             return false;
+            if (samplers.TryFindIdAndInsertIndex(1, out id, out index))
+            {
+                newSampler.Id = id;
+                samplers.Insert(index, newSampler);
+                return true;
+            }
+
+            return false;
         }
-        
+
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static byte AddWithId<T>(this DynamicBuffer<T> samplers, T newSampler) where T : struct, IElementWithId
         {
@@ -30,14 +32,16 @@ namespace DMotion
             return id;
         }
 
-        internal static bool TryFindIdAndInsertIndex<T>(this DynamicBuffer<T> samplers, byte reserveCount, out byte id, out int insertIndex) where T : struct, IElementWithId
+        internal static bool TryFindIdAndInsertIndex<T>(this DynamicBuffer<T> samplers, byte reserveCount, out byte id,
+            out int insertIndex) where T : struct, IElementWithId
         {
             //we assume the list is always sorted (should be true if Id always increments  from 0 to 128 and loops back)
             //on the loop back case, we add after the first element for which we can ensure reserveCount
             Assert.IsTrue(samplers.Length + reserveCount < MaxSamplersCount, "No support for more than 128 clips");
-            
+
             //sanity check
-            Assert.IsTrue(reserveCount <= MaxReserveCount, "Reserve count too high. Why are you trying to allocate so many contiguous clips?");
+            Assert.IsTrue(reserveCount <= MaxReserveCount,
+                "Reserve count too high. Why are you trying to allocate so many contiguous clips?");
 
             if (samplers.Length == 0)
             {
@@ -51,11 +55,11 @@ namespace DMotion
             if (idWithReserveCount < MaxSamplersCount)
             {
                 //impossible to overflow
-                id = (byte) (last.Id + 1);
+                id = (byte)(last.Id + 1);
                 insertIndex = samplers.Length;
                 return true;
             }
-            
+
             //possible loopback case
             //our last samplers had the max id, but there is no one else in the list. Give id 0
             if (samplers.Length == 1)
@@ -64,7 +68,7 @@ namespace DMotion
                 insertIndex = samplers.Length;
                 return true;
             }
-            
+
             //find first sampler for which we can have reserveCount contiguous indexes
             for (var i = 0; i < samplers.Length - 1; i++)
             {
@@ -73,12 +77,12 @@ namespace DMotion
 
                 if (next.Id - current.Id > reserveCount)
                 {
-                    id = (byte) (current.Id + 1);
+                    id = (byte)(current.Id + 1);
                     insertIndex = i + 1;
                     return true;
                 }
             }
-            
+
             // From my current understand, we can only be here if if (it's possible I'm wrong though): 
             // 1 - reserveCount is massive (asserted above), or 2 - we managed to get a very fragmented id space
             // We don't handle this case (it's not reasonable), so let's scream
@@ -87,7 +91,7 @@ namespace DMotion
             insertIndex = -1;
             return false;
         }
-        
+
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static bool RemoveWithId<T>(this DynamicBuffer<T> samplers, byte id) where T : struct, IElementWithId
         {
@@ -96,7 +100,8 @@ namespace DMotion
 
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool RemoveRangeWithId<T>(this DynamicBuffer<T> samplers, byte id, byte count) where T : struct, IElementWithId
+        internal static bool RemoveRangeWithId<T>(this DynamicBuffer<T> samplers, byte id, byte count)
+            where T : struct, IElementWithId
         {
             var index = samplers.IdToIndex(id);
             var exists = index >= 0;
@@ -118,6 +123,7 @@ namespace DMotion
                     return i;
                 }
             }
+
             return -1;
         }
 
@@ -126,9 +132,10 @@ namespace DMotion
         {
             return samplers.IdToIndex(id) >= 0;
         }
-        
+
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        internal static bool TryGetWithId<T>(this DynamicBuffer<T> samplers, byte id, out T element) where T : struct, IElementWithId
+        internal static bool TryGetWithId<T>(this DynamicBuffer<T> samplers, byte id, out T element)
+            where T : struct, IElementWithId
         {
             var index = samplers.IdToIndex(id);
             if (index >= 0)
@@ -144,7 +151,7 @@ namespace DMotion
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         internal static T GetWithId<T>(this DynamicBuffer<T> elements, byte id) where T : struct, IElementWithId
         {
-            var success= elements.TryGetWithId(id, out var e);
+            var success = elements.TryGetWithId(id, out var e);
             Assert.IsTrue(success);
             return e;
         }
